# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Team Photo Album API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      description: |
        Returns a presigned GCS URL for direct upload. The client sends JSON
        metadata here, then uploads the file directly to the returned URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          description: Missing or invalid required fields.
        '500':
          description: Failed to generate upload URL.
  /collections:
    get:
      operationId: listCollections
      tags: [collections]
      summary: List all collections
      responses:
        "200":
          description: List of collections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CollectionSummary"
        "401":
          description: Unauthorized
    post:
      operationId: createCollection
      tags: [collections]
      summary: Create a new collection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectionInput"
      responses:
        "201":
          description: Created collection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
        "401":
          description: Unauthorized
  /collections/{id}:
    get:
      operationId: getCollection
      tags: [collections]
      summary: Get a single collection with its photos
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Collection with photos
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
        "404":
          description: Not found
    patch:
      operationId: updateCollection
      tags: [collections]
      summary: Update a collection (owner or admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectionUpdate"
      responses:
        "200":
          description: Updated collection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
    delete:
      operationId: deleteCollection
      tags: [collections]
      summary: Delete a collection (owner or admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
  /collections/{id}/photos:
    post:
      operationId: addPhotoToCollection
      tags: [collections]
      summary: Add a photo to a collection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectionPhotoInput"
      responses:
        "204":
          description: Added
        "401":
          description: Unauthorized
        "404":
          description: Not found
  /collections/{id}/photos/{photoId}:
    delete:
      operationId: removePhotoFromCollection
      tags: [collections]
      summary: Remove a photo from a collection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: photoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
        "401":
          description: Unauthorized
        "404":
          description: Not found
  /collections/{id}/cover:
    patch:
      operationId: setCollectionCover
      tags: [collections]
      summary: Set the cover photo for a collection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CollectionCoverUpdate"
      responses:
        "200":
          description: Updated collection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Collection"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
  /collections/{id}/generate-keywords:
    post:
      operationId: generateCollectionKeywords
      tags: [collections]
      summary: Trigger AI keyword generation for a smart collection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Keywords generated and saved
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
...
components:
  schemas:
...
    CollectionCoverUpdate:
      type: object
      required: [photoId]
      properties:
        photoId:
          type: integer
    Photo:
      type: object
      required: [id, albumId, uploaderId, url, createdAt, ratingCount, isHidden]
      properties:
        id:
          type: integer
        albumId:
          type: integer
        albumTitle:
          type: ["string", "null"]
        uploaderId:
          type: integer
        storageKey:
          type: ["string", "null"]
        thumbnailKey:
          type: ["string", "null"]
        url:
          type: string
        filename:
          type: ["string", "null"]
        filesize:
          type: ["integer", "null"]
        takenAt:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        isHidden:
          type: boolean
        averageRating:
          type: ["number", "null"]
        ratingCount:
          type: integer
        myRating:
          type: ["integer", "null"]
        photoCollections:
          type: array
          items:
            $ref: "#/components/schemas/CollectionSummary"
        aiDescription:
          type: ["string", "null"]
        latestAiStatus:
          type: ["string", "null"]
          enum: ["success", "skipped", "failed", null]
          description: Status of the most recent AI analysis event for this photo
        suggestedCollections:
          type: array
          items:
            $ref: "#/components/schemas/SuggestedCollection"
        suggestedNewCollections:
          type: array
          items:
            $ref: "#/components/schemas/SuggestedNewCollection"
        ratings:
          type: array
          items:
            $ref: "#/components/schemas/PhotoRating"
...
    PhotoRating:
      type: object
      required: [userId, score, createdAt]
      properties:
        userId:
          type: integer
        userName:
          type: ["string", "null"]
        score:
          type: integer
        createdAt:
          type: string
          format: date-time
    CollectionSummary:
      type: object
      required: [id, title, createdById, createdAt, photoCount]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        createdById:
          type: integer
        photoCount:
          type: integer
        coverPhotoId:
          type: ["integer", "null"]
        coverPhotoUrl:
          type: ["string", "null"]
        coverPhotoThumbnailKey:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
    Collection:
      type: object
      required: [id, title, createdById, createdAt, photoCount]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        createdById:
          type: integer
        photoCount:
          type: integer
        coverPhotoId:
          type: ["integer", "null"]
        coverPhotoUrl:
          type: ["string", "null"]
        aiKeywords:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        photos:
          type: array
          items:
            $ref: "#/components/schemas/Photo"
    CollectionInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
    CollectionUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: ["string", "null"]
        coverPhotoId:
          type: ["integer", "null"]
        aiKeywords:
          type: ["string", "null"]
    CollectionPhotoInput:
      type: object
      required: [photoId]
      properties:
        photoId:
          type: integer
...
    SuggestedCollection:
      type: object
      required: [id, title]
      properties:
        id:
          type: integer
        title:
          type: string
    SuggestedNewCollection:
      type: object
      required: [id, suggestedName]
      properties:
        id:
          type: integer
        suggestedName:
          type: string
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
          description: Original file name.
        size:
          type: integer
          minimum: 1
          description: File size in bytes.
        contentType:
          type: string
          minLength: 1
          description: MIME type of the file (e.g. `image/jpeg`).
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          description: Presigned URL for PUT upload. Absolute in production; may be a relative proxy path (e.g. `/gcs/...`) in local dev.
        objectPath:
          type: string
          description: Normalized object path (e.g. `/objects/uploads/uuid`). Store this in your database.
        metadata:
          $ref: '#/components/schemas/UploadUrlRequest'
...