# lib/api-zod/src/generated/types/collection.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Team Photo Album API
 * OpenAPI spec version: 0.1.0
 */
import type { Photo } from "./photo";

export interface Collection {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  createdById: number;
  photoCount: number;
  /** @nullable */
  coverPhotoId?: number | null;
  /** @nullable */
  coverPhotoUrl?: string | null;
  /** @nullable */
  aiKeywords?: string | null;
  tags?: string[];
  createdAt: Date;
  photos?: Photo[];
}